+++ CHANGELOG.md
@@ -5,7 +5,10 @@
 
 ### Features
 1. [#257](https://github.com/influxdata/influxdb-client-csharp/pull/257): Add `PingService` to check status of OSS and Cloud instance
-2. [#260](https://github.com/influxdata/influxdb-client-csharp/pull/260): Changed `internal` to `public` visiblity of `InfluxDBClientOptions.Builder.ConnectionString`
+1. [#260](https://github.com/influxdata/influxdb-client-csharp/pull/260): Changed `internal` to `public` visibility of `InfluxDBClientOptions.Builder.ConnectionString`
+
+### CI
+1. [#264](https://github.com/influxdata/influxdb-client-csharp/pull/264): Add build for `dotnet6`
 
 ### Bug Fixes
 1. [#262](https://github.com/influxdata/influxdb-client-csharp/issues/262): InfluxDB 2.1 Incompatibility with Session Cookie
@@ -140,7 +143,6 @@ This release introduces a support for new InfluxDB OSS API definitions - [oss.ym
 
 ### CI
 1. [#182](https://github.com/influxdata/influxdb-client-csharp/pull/182): Add build for `dotnet5`, Fix code coverage report
-
 
 ## 1.16.0 [2021-03-05]