+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 3.2.0 [unreleased]
+## 3.2.0 [2021-11-26]
 
 ### Deprecates
 - `InfluxDBClient.HealthAsync()`: instead use `InfluxDBClient.PingAsync()`